requirements.txt
selenium==4.18.1
pytest==8.0.2
pytest-xdist==3.5.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
python-dotenv==1.0.1
requests==2.31.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
execnet==2.1.1
Faker==36.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest-rerunfailures==15.0
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0 